FEAT: project member connection

+++ jjapra/src/main/java/jjapra/app/controller/ProjectController.java
@@ -25,17 +25,14 @@ public class ProjectController {
 
     @GetMapping("")
     public ResponseEntity<List<ProjectMember>> getProjects(HttpSession session) {
-        Object loggedInUser = session.getAttribute("loggedInUser");
-        if (loggedInUser == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
-        }
-
+        Member loggedInUser = (Member) session.getAttribute("loggedInUser");
         List<ProjectMember> projects = projectMemberService.findByMemberId(((Member) loggedInUser).getId());
         return ResponseEntity.status(HttpStatus.OK).body(projects);
     }
 
     @PostMapping("")
     public ResponseEntity<Project> addProject(@RequestBody AddProjectRequest request, HttpSession session) {
+        // admin인 경우에만 프로젝트 생성할 수 있도록 코드 추가하기
         if (request.getTitle().isEmpty()) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
         }
@@ -47,7 +44,7 @@ public class ProjectController {
         ProjectMember projectMember = ProjectMember.builder()
                 .project(savedProject)
                 .member((Member) session.getAttribute("loggedInUser"))
-                .role(Role.PL)
+                .role(Role.ADMIN)
                 .build();
         projectMemberService.save(projectMember);
 
@@ -55,25 +52,41 @@ public class ProjectController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Project> getProject(@PathVariable("id") Integer id) {
-        Project project = projectService.findById(id);
-        if (project == null) {
+    public ResponseEntity<Project> getProject(@PathVariable("id") Integer id, HttpSession session) {
+        List<ProjectMember> projectMembers = projectMemberService.findByProjectId(id);
+        if (projectMembers.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
         }
-        return ResponseEntity.status(HttpStatus.OK).body(project);
+        if (projectMembers.stream().noneMatch(pm -> pm.getMember().getId().equals((
+                (Member) session.getAttribute("loggedInUser")).getId()))) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).body(projectMembers.get(0).getProject());
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Project> updateProject(@PathVariable("id") Integer id, @RequestBody AddProjectRequest request) {
-        Project updatedProject = projectService.update(id, request);
-        if (updatedProject == null) {
+    public ResponseEntity<Project> updateProject(@PathVariable("id") Integer id, @RequestBody AddProjectRequest request, HttpSession session) {
+        List<ProjectMember> projectMembers = projectMemberService.findByProjectId(id);
+        if (projectMembers.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
         }
+        if (projectMembers.stream().noneMatch(pm -> pm.getMember().getId().equals((
+                (Member) session.getAttribute("loggedInUser")).getId()))) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+
+        Project updatedProject = projectService.update(id, request);
         return ResponseEntity.status(HttpStatus.OK).body(updatedProject);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Project> deleteProject(@PathVariable("id") Integer id) {
+        List<ProjectMember> projectMemberList = projectMemberService.findByProjectId(id);
+        if (projectMemberList.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+        projectMemberList.forEach(projectMemberService::delete);
         return projectService.deleteProject(id);
     }
 }
